Anagram

Description

Core anagram toolkit. The main entry points are anagram, anagramAny, anagramMin, anagramFull, and anagramExact. Histograms are the internal form for anagram queries, and are a count of how many of each letter is present in the query.

Examples

>>> anagram ukacd "eoynkm"
["monkey", "mekon", "money", ... "yon"]
>>> anagramAny ukacd "eoynkm"
["monkey", "mekon", "money", ... "me", "mo", "my", ... "n", "o", "y"]
>>> anagramFull ukacd "eoynkm"
["monkey"]
>>> anagramMin ukacd "eoynkm" 5
["monkey", "mekon", "money"]
>>> anagramExact ukacd 5 "eoynkm"
["mekon", "money"]
>>> anagramFull ukacd "?eoynk"
["donkey","monkey","orkney","unyoke","yonker"]

We can also use the Histogram type directly, if that happens to be convenient:

>>> Histogram 9 (replicate 26 1)
histogramFromPairs 9 [('a',1),('b',1),('c',1),('d',1),('e',1),('f',1),('g',1),('h',1),('i',1),
  ('j',1),('k',1),('l',1),('m',1),('n',1),('o',1),('p',1),('q',1),('r',1),('s',1),('t',1),
  ('u',1),('v',1),('w',1),('x',1),('y',1),('z',1)]
>>> anagramFull onelook (Histogram 9 (replicate 26 1))
["thequickbrownfoxjumpsoverthelazydog"]

Documentation

class Anagrammable a where

A typeclass for things that can be treated as anagram queries.

Currently, there is a String instance (so you can use bare strings; also known as [Char]) and a Histogram instance, to allow lower-level access through the same functions.

Minimal complete definition

getHistogram

Methods

getHistogram :: a -> Histogram

Get the query in the form of a Histogram, to actually use to access an anagram dictionary.

Main entry points

anagram :: Anagrammable a => AnagramDictionary -> a -> [String]

Basic anagram. Constrains the set of returned anagrams to use between half and all of the letters in the query.

anagramAny :: Anagrammable a => AnagramDictionary -> a -> [String]

As anagram, but from all to none of the letters in the query.

anagramFull :: Anagrammable a => AnagramDictionary -> a -> [String]

Anagram all the letters in the query.

anagramFull dict query = anagramMin dict query (length query)

anagramMin :: Anagrammable a => AnagramDictionary -> Int -> a -> [String]

Anagram at-least-k letters from the query.

anagram dict query = anagramMin dict (length query / 2) query

anagramAny dict query = anagramMin dict 0 query 0

anagramExact :: Anagrammable a => AnagramDictionary -> Int -> a -> [String]

Anagram exactly k letters from the query.

all (\a->length a == k) (anagramExact dict k query)

Lower-level programmatic interface

data Histogram

The Histogram type; we implement anagram queries by converting the string to one of these, and querying the database for matches.

Constructors

Histogram

Fields

blankCount :: Int

We differentiate internally between real letters and wildcards; the underlying code has to deal with them fully differently.

fromHistogram :: [Int]

the list of counts; the order is the same order as anagramSymbols.

anagramSymbols :: [Char]

The list of symbols we support anagramming; anything not in this list is simply dropped. We use a finite list so Histogram and AnagramDictionary don't need to store actual characters.

getHistogram :: Anagrammable a => a -> Histogram

Get the query in the form of a Histogram, to actually use to access an anagram dictionary.

subHistogram :: Histogram -> Histogram -> Histogram

Subtract a histogram from a histogram. "aabc" - "ab" = "ac"

Dictionaries

newtype AnagramDictionary

A dictionary structure for querying anagrams.

Internally, this is just the same sort of trie structure as CrosswordDictionary, but instead of storing actual characters we treat the members as Word8s of letter counts, followed by the actual word. This allows us to eliminate common tails among the given words essentially for free.

Constructors

AnagramDictionary Node

buildAnagramDictionary :: [[Char]] -> AnagramDictionary

Construct an anagram dictionary from a raw list of strings. Gets very slow on large dictionaries, due to the time to construct the trie.

fromAnagramDictionary :: AnagramDictionary -> Node

Get the raw trie from an AnagramDictionary.