AnswerChecker

Description

Documentation

checkAnswer :: String -> String -> Correctness

Check an answer, assuming that it follows exactly the puzzle-solution CSS class and uppercase pattern.

checkAnswerFuzzy :: String -> String -> ByteString

Pattern match against the given answer, in uppercase, on it's own within any pair of HTML tags. Can leak answer information, use only if confident and checkAnswer doesn't work.

data Correctness

Simple "are you correct?" type, with a Show instance.

Constructors

CORRECT

INCORRECT

NONEXISTANT