Cipher

Description

Currently all that's here are rot n-related ciphers, but more to be added.

Documentation

rot :: Int -> [Char] -> [Char]

rotN cipher.

rotAll :: [Char] -> [[Char]]

Brute-force a rotN cipher, by showing a list of each rotation. If there's a correct answer, it should be easy enough to pick out by eye.