Crossword

Description

Core regex-like dictionary search; we implement a simple glob-like pattern, and nondeterministic retrieval from a trie structure to minimize cost.

Examples

>>> crossword onelook "j???l??p??a?d"
["jeanlucpicard", ...]

Documentation

class CrosswordQuery a

A typeclass of types we can use as a crossword query.

finalizeQuery and finalizeQueryStraight are probably going to merge into a single getQueryParts function. The distinction now is that finalizeQuery produces a query appropriate for a BidirectionalDictionary, while finalizeQueryStraight produces a query appropriate for a ForwardDictionary.

Minimal complete definition

finalizeQuery, finalizeQueryStraight

Entry points

crossword :: CrosswordQuery a => CrosswordDictionary -> a -> [String]

Run the given query against a crossword dictionary. This is the primary entry point for this module.

crosswordSubseq :: CrosswordDictionary -> [Char] -> [[Char]]

Treat the input as a raw string, and query for subsequences of that string in the dictionary. Does not have metacharacters, and essentially wraps every letter in an individual Opt.

Programmatic interface

data QueryPart

Constructors

Literal String
A literal string of characters. This could equivalently be written with as a Literal Char type, but we chose to simplify the QueryPart level of the representation.

Glob
Matches zero-or-more arbitrary characters. Notably, this is not the kleene star; in regex terms it's .*. * is parsed to Glob.

Dot
Exactly one arbitrary character. Named after regex dot, but we parse ? to it following shell and crossword-solver conventions.

Charset [Char]
Character sets; matches exactly one of any of the given characters. Not available from the string-to-query parser.

Star QueryPart
Kleene star. Not available from the string-to-query parser.

Opt String
FIXME: what does this do again?

Dictionaries

data CrosswordDictionary

A crossword dictionary.

Constructors

BidirectionalDictionary Node
A dictionary that has been munged to allow starting in the middle of the word; see optimizeInitialLiteral in the source for details.

ForwardDictionary Node
A dictionary that is a simple trie of the words in the dictionary.

buildDict :: [String] -> Node

buildDictAscSplit :: [String] -> Node

buildDictUnidir :: [String] -> Node

getCrosswordDAWG :: CrosswordDictionary -> Node