Dictionary

Description

We use a simple return type polymorphism mechanism to select the appropriate data structure for the query, so we have things like ukacd :: (DictionaryContext a) => a where a can be either CrosswordDictionary or AnagramDictionary. The definition for ukacd is then something like

ukacd = fromDictionary $ Dictionary (AnagramDictionary $ fromFile "ukacd-anagram.dawg") (Crossword.BidirectionalDictionary $ fromFile "ukacd-cross.dawg")

(not literally that, there are some unsafePerformIO (We know that these files are effectively large constants), "where is this file installed", etc. steps, but that is roughly the correct intuition.)

Documentation

data Dictionary

Effectively, a Dictionary object consists of a copy of the dictionary structured for anagrams and the same dictionary structured for crossword queries. If we find that there are more data structures that are useful and we want all the dictionaries in, we can add them here (and add an appropriate DictionaryContext instance).

Constructors

Dictionary

Fields

anaDict :: AnagramDictionary

crossDict :: CrosswordDictionary

class DictionaryContext a where

A context that we might want to have a specialized dictionary representation for.

Minimal complete definition

fromDictionary

Methods

fromDictionary :: Dictionary -> a

fromDictionary :: DictionaryContext a => Dictionary -> a